go-webserver: move group fetcher function into preparer function
