+++ go/src/koding/go-webserver/userinfo.go
@@ -40,13 +40,43 @@ type UserInfo struct {
 	Group *models.Group
 }
 
+func getGroup(r *http.Request) (*models.Group, error) {
+	c, err := r.Cookie("groupName")
+	if err != nil && err != http.ErrNoCookie {
+		return nil, err
+	}
+
+	// initial group name
+	groupName := ""
+
+	// try to get cookie value,
+	// TODO ~ when we fully implement the feature, be more cautious here
+	if c != nil && c.Value != "" {
+		groupName = c.Value
+	} else {
+		Log.Error("couldnt find groupname, setting koding as group for now")
+		groupName = "koding"
+	}
+
+	// TODO implement caching here
+	group, err := modelhelper.GetGroup(groupName)
+	if err != nil {
+		return nil, err
+	}
+
+	return group, nil
+}
+
 // fetchUseInfo fetches different user models and returns UserInfo.  It first
 // fetches `clientId` cookie and if it exists, it fetches JSession and other
 // info.
-func prepareUserInfo(w http.ResponseWriter, r *http.Request, group *models.Group) (*UserInfo, error) {
-	subdomain := r.Header.Get("Referer")
-	fmt.Println("r.Header-->", r.Header)
-	fmt.Println("subdomain-->", subdomain)
+func prepareUserInfo(w http.ResponseWriter, r *http.Request) (*UserInfo, error) {
+	group, err := getGroup(r)
+	if err != nil {
+		Log.Error("err while getting group %s", err.Error())
+		return nil, err
+	}
+
 	cookie, err := getCookie(w, r)
 	if err != nil {
 		return nil, err
